requirements.txt
socket
requests
asyncio
dnspython
beautifulsoup4
colorama
rich